316-Fix-BasicCarMaintenanceUITests
This is my first ever open source contribution, and first time working with XCTest, so naturally I have many questions:

The error I was getting said that it was failing to scroll. I thought it may have something to do with the 'Thanks for using this app' popup so I found a way to dismiss it.

Dismissing the popup allowed the test to pass. But the test isn't actually asserting anything. Is that right for a UI test? A successful tap of the 'add' button is a successful test? Should I be asserting that a new vehicle has been added?

It looks like I set up firebase correctly for when I run the app, but the data doesn't go away each time I run the tests, have I missed a step or is that working as expected?

Should I edit the test to fill in the rest of the inputs?

Thanks